unnecessary boxing to Long removed

+++ visualvm/sampler/src/org/graalvm/visualvm/sampler/memory/ThreadsMemoryView.java
@@ -503,10 +503,10 @@ final class ThreadsMemoryView extends JPanel {
             if (columnIndex == 0) {
                 return threads.get(rowIndex).getThreadName();
             } else if (columnIndex == 1) {
-                return allocatedBytes.get(rowIndex).longValue();
+                return allocatedBytes.get(rowIndex);
             } else if (columnIndex == 2) {
                 return allocatedBytesPerSec.isEmpty() ? 0 :
-                       allocatedBytesPerSec.get(rowIndex).longValue();
+                       allocatedBytesPerSec.get(rowIndex);
             }
 
             return null;
